backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
scanpy>=1.10.0
matplotlib>=3.8.0
leidenalg>=0.10.0
python-multipart>=0.0.9
anndata>=0.10.0
scipy>=1.12.0
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.26.0
google-genai>=1.0.0
